Command parser is now mostly functional, still need to fix several things
