+++ minishell.h
@@ -6,7 +6,7 @@
 /*   By: paulorod <paulorod@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/20 15:31:38 by paulorod          #+#    #+#             */
-/*   Updated: 2023/08/29 14:30:04 by paulorod         ###   ########.fr       */
+/*   Updated: 2023/08/30 16:03:33 by paulorod         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,6 +29,8 @@ void	clear_env_list(t_list **env_list);
 void	register_signals(void);
 void	free_cmd(t_cmd *cmd);
 void	clear_paths(char **paths);
+void	insert_end(t_cmd **head, char *cmd);
+void	insert_front(t_cmd **head, char *cmd);
 int		print_fd(char *error, char fd, char *name);
 int		create_command_process(t_cmd *cmd, char **env);
 char	**alloc_cmd(char *command);
@@ -37,5 +39,7 @@ char	*get_full_path(char **paths, char *command);
 char	*search_command_path(char *command);
 char	**create_cmd_tokens(char *command, t_shell *shell);
 char	*ft_getenv(const char *name, char ***_env);
+
+t_cmd	*create_cmd_list(char **tokens, t_shell *shell);
 
 #endif
