+++ srcs/utils/commands.c
@@ -6,7 +6,7 @@
 /*   By: paulorod <paulorod@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/10 16:16:56 by paulorod          #+#    #+#             */
-/*   Updated: 2023/08/30 13:43:24 by paulorod         ###   ########.fr       */
+/*   Updated: 2023/08/30 16:22:56 by paulorod         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,8 +54,8 @@ static char	*handle_envs(char *command, t_shell *shell)
 	return (command);
 }
 
-/*Create command struct*/
-//TODO change to split by delimiter
+/* OLD CODE
+Create command struct
 char	**old_create_cmd(char *_cmd, t_shell *shell)
 {
 	char			**cmd;
@@ -83,152 +83,42 @@ char	**old_create_cmd(char *_cmd, t_shell *shell)
 		i++;
 	}
 	return (cmd);
+}*/
+
+/*Get command type indentifier*/
+enum e_identifiers	get_cmd_type(char *token)
+{
+	if (ft_strlen(token) == 1)
+	{
+		if (*token == '|')
+			return (_pipe);
+		if (*token == '>')
+			return (greater);
+		if (*token == '<')
+			return (lesser);
+	}
+	else
+	{
+		if (*token == '>')
+			return (output);
+		if (*token == '<')
+			return (input);
+	}
+	return (_command);
 }
 
-/*Check if character is between quotes*/
-bool	in_quotes(char c)
+/*Create empty command struct for pipes and redirections*/
+t_cmd	*create_token_cmd(char *token)
 {
-	static bool	d_quote = false;
-	static bool	s_quote = false;
+	t_cmd	*command;
 
-	if (c == '"' && !s_quote)
-		d_quote = !d_quote;
-	if (c == '\'' && !d_quote)
-		s_quote = !s_quote;
-	return (d_quote || s_quote);
-}
-
-/*Check for unclosed quotes*/
-char	*check_unclosed_quotes(char *command)
-{
-	char	*rl_tmp;
-	char	*tmp;
-	char	*promt;
-	int		i;
-
-	if (in_quotes(0))
-	{
-		if (!in_quotes('"'))
-			promt = "dquote>";
-		else if (!in_quotes('\''))
-			promt = "quote>";
-		rl_tmp = readline(promt);
-		tmp = ft_strjoin(command, rl_tmp);
-		free(command);
-		free(rl_tmp);
-		command = tmp;
-	}
-	i = 0;
-	while (command[i])
-	{
-		in_quotes(command[i]);
-		i++;
-	}
-	if (in_quotes(0))
-		command = check_unclosed_quotes(command);
+	command = ft_calloc(sizeof(t_cmd), 1);
+	command->indentifier = (enum e_identifiers)get_cmd_type(token);
 	return (command);
 }
 
-/*Prepare string for token generation*/
-char	*prepare_string(char *command)
-{
-	int		i;
-
-	i = 0;
-	while (command[++i])
-	{
-		if (in_quotes(command[i]))
-			continue ;
-		if (command[i] == ' ')
-			command[i] = '\1';
-	}
-	command = check_unclosed_quotes(command);
-	return (command);
-}
-
-/*Check if position "i" in string is a special character*/
-bool	is_special_char(char *str, int i, int *end)
-{
-	if (str[i] == '|')
-		return (true);
-	if (str[i] == '<' && str[i + 1] != '<' && str[i - 1] != '<')
-		return (true);
-	if (str[i] == '>' && str[i + 1] != '>' && str[i - 1] != '>')
-		return (true);
-	if (str[i] == '<' && str[i + 1] == '<')
-	{
-		if (end)
-			*end = i + 1;
-		return (true);
-	}
-	if (str[i] == '>' && str[i + 1] == '>')
-	{
-		if (end)
-			*end = i + 1;
-		return (true);
-	}
-	return (false);
-}
-
-/*Copy old string and add separator between special characters*/
-char	*add_separators(char *command, int start, int end)
-{
-	int		j;
-	char	*new_command;
-
-	new_command = ft_calloc(sizeof(char), ft_strlen(command) + 2);
-	j = 0;
-	while (command[j] && j < start)
-	{
-		new_command[j] = command[j];
-		j++;
-	}
-	new_command[j++] = '\1';
-	while (j <= end + 1)
-	{
-		new_command[j] = command[j - 1];
-		j++;
-	}
-	new_command[j++] = '\1';
-	while (command[j - 2])
-	{
-		new_command[j] = command[j - 2];
-		j++;
-	}
-	new_command[j] = '\0';
-	return (new_command);
-}
-
-/*Add separator (\1) between special characters*/
-char	*separate_special_chars(char *command)
-{
-	int	i;
-	int	end;
-
-	i = 0;
-	end = 0;
-	while (command[i])
-	{
-		end = i;
-		if (!in_quotes(command[i]) && is_special_char(command, i, &end))
-		{
-			if (command[i + 1] == '<' || command[i + 1] == '>')
-			{
-				if (command[i - 1] != '\1' || command[i + 2] != '\1')
-					command = add_separators(command, i, end);
-			}
-			else
-			{
-				if (command[i - 1] != '\1' || command[i + 1] != '\1')
-					command = add_separators(command, i, end);
-			}
-		}
-		i++;
-	}
-	return (command);
-}
-
-t_cmd	*create_cmd_list(char **tokens)
+/*Create linked list of command structs*/
+t_cmd	*create_cmd_list(char **tokens, t_shell *shell)
 {
 	int		i;
 	int		j;
@@ -241,25 +131,17 @@ t_cmd	*create_cmd_list(char **tokens)
 	while (tokens[i])
 	{
 		if (!is_special_char(tokens[i], 0, NULL))
-			command->cmd[j++] = tokens[i];
+		{
+			command->indentifier = _command;
+			command->cmd[j++] = handle_envs(tokens[i], shell);
+		}
 		else
-			command->next = create_cmd_list(&tokens[i + 1]);
+		{
+			command->next = create_token_cmd(tokens[i]);
+			command->next->next = create_cmd_list(&tokens[i + 1], shell);
+		}
 		i++;
 	}
+	command->cmd[j] = NULL;
 	return (command);
 }
-
-/*Split command into tokens*/
-char	**create_cmd_tokens(char *command, t_shell *shell)
-{
-	char	**cmd;
-
-	(void)shell;
-	command = prepare_string(command);
-	command = separate_special_chars(command);
-	cmd = ft_split(command, '\1');
-	t_cmd *test = create_cmd_list(cmd);
-	(void)test;
-	return (cmd);
-}
-
